requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.6
Django==4.1.5
django-admin==2.0.1
django-crispy-forms==1.13.0
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-six==1.0.4
phonenumbers==8.13.4
protobuf==3.20.1
PyMySQL==1.0.2
pytz==2022.7
screen==1.0.1
sqlparse==0.4.3
xlwt==1.3.0
